refactor(i've added check on error for promises): i've added a check on error for promises
i've added a check on error for promises. I think that this check need because if will be an error
with read from mongo an entity then this check will say that has occured an error when a user wanted
to read an entity from mongo

BREAKING CHANGE: i've added a check on error for promises. I think that this check need because if
will be an error iwth read from mongo an entity then this check will say that has occured an error
when a user wanted to read an entity from mongo

fix errors with mongo crud